requirements.txt
PyAutoGUI==0.9.53
opencv_python==4.5.3.56
numpy==1.19.5
Pillow==8.3.1
PySide2==5.15.2
